Try to make stork faster
