+++ stork/_version.py
@@ -1,1 +1,1 @@
-__version__ = '3.1.0'
+__version__ = '3.2.0'
