+++ stork/update_databricks_library.py
@@ -11,137 +11,9 @@ import requests
 from configparser import NoOptionError
 from simplejson.errors import JSONDecodeError
 
+from .api_error import APIError
 from .configure import _load_config, CFG_FILE, PROFILE
-
-
-class APIError(Exception):
-    """
-    exception to handle unifying two generations of API error responses
-     from Databricks
-    """
-    def __init__(self, response):
-        Exception.__init__(self, response)
-        try:
-            res_body = response.json()
-        except JSONDecodeError:
-            self.code = 'http {}'.format(response.status_code)
-            # non-json error message, didn't bother parsing neatly
-            self.message = response.text
-        else:
-            if 'error_code' in res_body.keys():
-                self.code = res_body['error_code']
-                self.message = res_body['message']
-            else:
-                self.code = 'http {}'.format(response.status_code)
-                self.message = res_body['error']
-
-    def __str__(self):
-        return '{}: {}'.format(self.code, self.message)
-
-
-class FileNameError(Exception):
-    """
-    exception to handle when filename is not of correct pattern
-    """
-    def __init__(self, filename):
-        Exception.__init__(
-            self,
-            'Filename \'{}\' was not correct pattern'.format(filename)
-        )
-        self.filename = filename
-
-
-class FileNameMatch(object):
-    """
-    Matches eggs or jars for both released and snapshot versions
-
-    Supported Patterns:
-      new_library-1.0.0-py3.6.egg
-      new_library-1.0.0-SNAPSHOT-py3.6.egg
-      new_library-1.0.0-SNAPSHOT-my-branch-py3.6.egg
-
-      new_library-1.0.0.egg
-      new_library-1.0.0-SNAPSHOT.egg
-      new_library-1.0.0-SNAPSHOT-my-branch.egg
-
-      new_library-1.0.0.jar
-      new_library-1.0.0-SNAPSHOT.jar
-      new_library-1.0.0-SNAPSHOT-my-branch.jar
-
-    Parameters
-    ----------
-    library_name: string
-        base name of library (e.g. 'test_library')
-    version: string
-        version of library (e.g. '1.0.0')
-
-    """
-    file_pattern = (
-        r'([a-zA-Z0-9-\._]+)-((\d+)\.(\d+\.\d+)'
-        r'(?:-SNAPSHOT(?:[a-zA-Z_\-\.]+)?)?)(?:-py.+)?\.(egg|jar)'
-    )
-
-    def __init__(self, filename):
-        match = re.match(FileNameMatch.file_pattern, filename)
-        try:
-            self.filename = filename
-            self.library_name = match.group(1)
-            self.version = match.group(2)
-            self.major_version = match.group(3)
-            self.minor_version = match.group(4)
-            self.suffix = match.group(5)
-            if self.suffix == 'jar':
-                self.lib_type = 'java-jar'
-            elif self.suffix == 'egg':
-                self.lib_type = 'python-egg'
-        except (IndexError, AttributeError):
-            raise FileNameError(filename)
-
-    def __eq__(self, other):
-        if isinstance(other, self.__class__):
-            self_attrs = {k: v for k, v in vars(self).items()}
-            self_attrs.pop('filename')
-            other_attrs = {k: v for k, v in vars(other).items()}
-            other_attrs.pop('filename')
-            return self_attrs == other_attrs
-        else:
-            return False
-
-    def replace_version(self, other, logger):
-        """
-        True if self can safely replace other
-
-        based on version numbers only - snapshot and branch tags are ignored
-        """
-
-        if other.library_name != self.library_name:
-            logger.debug(
-                'not replacable: {} != {} ()'
-                .format(other.library_name, self.library_name)
-            )
-            return False
-        elif int(other.major_version) != int(self.major_version):
-            logger.debug(
-                'not replacable: {} != {} ({})'
-                .format(
-                    int(self.major_version),
-                    int(other.major_version),
-                    other.filename,
-                )
-            )
-            return False
-        elif float(other.minor_version) >= float(self.minor_version):
-            logger.debug(
-                'not replacable: {} >= {} ({})'
-                .format(
-                    other.minor_version,
-                    self.minor_version,
-                    other.filename,
-                )
-            )
-            return False
-        else:
-            return True
+from .file_name import FileNameError, FileNameMatch
 
 
 def load_library(filename, match, folder, token, host):
@@ -274,67 +146,61 @@ def get_library_mapping(logger, prod_folder, token, host):
         minor version, and id number
     """
     res = requests.get(
-        host + '/api/1.2/libraries/list',
+        host + f'/api/2.0/workspace/list?path={prod_folder}',
         auth=('token', token),
     )
     if res.status_code == 200:
-        library_list = res.json()
+        file_list = res.json()['objects']
         library_map = {}
         id_nums = {}
-        for library in library_list:
-            status_res = (
-                requests
-                .get(
-                    host + '/api/1.2/libraries/status?libraryId={}'
-                    .format(library['id']),
-                    auth=('token', token),
-                )
-            )
-            if status_res.status_code == 200:
-                library_info = status_res.json()
-                # only do any of this for libraries in the production folder
-                if library_info['folder'] != prod_folder:
-                    logger.debug(
-                        'excluded folder: {} in {}, not prod folder ({})'
-                        .format(
-                            library_info['name'],
-                            library_info['folder'],
-                            prod_folder,
-                        )
-                    )
-                    continue
-                if library_info['libType'] == 'python-egg':
-                    full_name = library_info['name'] + '.egg'
-                elif library_info['libType'] == 'java-jar':
-                    full_name = library_info['name'] + '.jar'
-                else:
-                    logger.debug(
-                        'excluded library type: {} is of libType {}, '
-                        'not jar or egg'
-                        .format(
-                            library_info['name'],
-                            library_info['libType'],
-                        )
-                    )
-                    continue
+        for file in file_list:
+            if file['object_type'] == 'LIBRARY':
+                library_id = file['object_id']
                 try:
-                    name_match = FileNameMatch(full_name)
-                    # map uri to name match object
-                    library_map[library_info['files'][0]] = name_match
-                    # map name to name match object and id number
-                    # we'll need the id number to clean up old libraries
-                    id_nums[library_info['name']] = {
-                        'name_match': name_match,
-                        'id_num': library_info['id'],
-                    }
-                except FileNameError:
-                    logger.debug(
-                        'FileNameError: {} file name is not parsable'
-                        .format(full_name)
+                    status_res = (
+                        requests
+                        .get(
+                            host + f'/api/1.2/libraries/status?libraryId={library_id}',
+                            auth=('token', token),
+                        )
                     )
-                    pass
-            else:
-                raise APIError(status_res)
+                except:
+                    logger.debug(f'Status request for library: {library_id} failed')
+                    continue
+                if status_res.status_code == 200:
+                    library_info = status_res.json()
+                    if library_info['libType'] == 'python-egg':
+                        full_name = library_info['name'] + '.egg'
+                    elif library_info['libType'] == 'java-jar':
+                        full_name = library_info['name'] + '.jar'
+                    else:
+                        logger.debug(
+                            'excluded library type: {} is of libType {}, '
+                            'not jar or egg'
+                            .format(
+                                library_info['name'],
+                                library_info['libType'],
+                            )
+                        )
+                        continue
+                    try:
+                        name_match = FileNameMatch(full_name)
+                        # map uri to name match object
+                        library_map[library_info['files'][0]] = name_match
+                        # map name to name match object and id number
+                        # we'll need the id number to clean up old libraries
+                        id_nums[library_info['name']] = {
+                            'name_match': name_match,
+                            'id_num': library_id,
+                        }
+                    except FileNameError:
+                        logger.debug(
+                            'FileNameError: {} file name is not parsable'
+                            .format(full_name)
+                        )
+                        pass
+                else:
+                    raise APIError(status_res)
         return library_map, id_nums
     else:
         raise APIError(res)
@@ -495,6 +361,7 @@ def update_databricks(logger, path, token, folder, update_jobs, cleanup):
                          ' to get set up')
     try:
         prod_folder = config.get(PROFILE, 'prod_folder')
+        prod_folder = '/Shared/DS_libraries'
     except NoOptionError:
         raise ValueError('no prod_folder provided: please run '
                          '`stork configure` to get set up')
@@ -537,7 +404,6 @@ def update_databricks(logger, path, token, folder, update_jobs, cleanup):
             )
         ][0]
         library_path = 'dbfs:/FileStore/jars/' + library_uri
-
         job_list = get_job_list(logger, match, library_map, token, host)
         logger.info(
             'current major version of library used by jobs: {}'
